refactor: types 폴더의 파일명 이름 개선
