+++ src/stores/Restaurant.ts
@@ -1,5 +1,5 @@
 import { atom } from 'recoil';
-import { RestaurantProps } from 'types/Restaurant';
+import { RestaurantProps } from 'types/restaurant';
 
 export const searchRestaurantListState = atom<RestaurantProps[]>({
   key: 'searchRestaurantList',
